Completed the route folder and moved everthing there
